refactor: 평가 모델 네이밍 수정
