+++ usw-suwiki-lecture/src/main/java/usw/suwiki/domain/lecture/model/Evaluation.java
@@ -1,11 +1,11 @@
-package usw.suwiki.domain.lecture.data;
+package usw.suwiki.domain.lecture.model;
 
 /**
  * EvaluatePost를 작성함으로 평가된 데이터 정보들로, 모듈 간의 데이터를 넘겨주는 용도로 사용된다.
  *
  * @author hejow
  */
-public record EvaluatedData(
+public record Evaluation(
   float totalAverage,
   float satisfaction,
   float honey,
