+++ usw-suwiki-post/usw-suwiki-evaluatepost/src/main/java/usw/suwiki/domain/evaluatepost/service/EvaluatePostService.java
@@ -11,7 +11,7 @@ import usw.suwiki.domain.evaluatepost.EvaluatePostQueryRepository;
 import usw.suwiki.domain.evaluatepost.EvaluatePostRepository;
 import usw.suwiki.domain.evaluatepost.dto.EvaluatePostRequest;
 import usw.suwiki.domain.evaluatepost.dto.EvaluatePostResponse;
-import usw.suwiki.domain.lecture.data.EvaluatedData;
+import usw.suwiki.domain.lecture.model.Evaluation;
 import usw.suwiki.domain.lecture.service.LectureService;
 import usw.suwiki.domain.report.model.Report;
 import usw.suwiki.domain.report.service.ReportService;
@@ -58,9 +58,9 @@ public class EvaluatePostService {
     }
 
     EvaluatePost evaluatePost = EvaluatePostMapper.toEntity(userId, lectureId, request);
-    EvaluatedData evaluatedData = EvaluatePostMapper.toEvaluatedData(evaluatePost.getLectureRating());
+    Evaluation evaluation = EvaluatePostMapper.toEvaluatedData(evaluatePost.getLectureRating());
 
-    lectureService.evaluate(lectureId, evaluatedData);
+    lectureService.evaluate(lectureId, evaluation);
     evaluatePostRepository.save(evaluatePost);
     userBusinessService.wroteEvaluation(userId);
   }
@@ -71,7 +71,7 @@ public class EvaluatePostService {
 
   public void update(Long evaluateId, EvaluatePostRequest.Update request) {
     EvaluatePost evaluatePost = loadEvaluatePostById(evaluateId);
-    EvaluatedData currentEvaluation = EvaluatePostMapper.toEvaluatedData(evaluatePost.getLectureRating());
+    Evaluation currentEvaluation = EvaluatePostMapper.toEvaluatedData(evaluatePost.getLectureRating());
 
     evaluatePost.update(
       evaluatePost.getContent(),
@@ -81,7 +81,7 @@ public class EvaluatePostService {
       EvaluatePostMapper.toRating(request)
     );
 
-    EvaluatedData updatedEvaluation = EvaluatePostMapper.toEvaluatedData(evaluatePost.getLectureRating());
+    Evaluation updatedEvaluation = EvaluatePostMapper.toEvaluatedData(evaluatePost.getLectureRating());
     lectureService.updateEvaluation(evaluatePost.getLectureId(), currentEvaluation, updatedEvaluation);
   }
